requirements.txt
firebase-admin>=6.0.0
python-dotenv>=0.19.0
feedparser>=6.0.8
newspaper3k>=0.2.8
nltk>=3.6.7
scikit-learn>=1.0.2
sentence-transformers>=2.2.2
python-dateutil>=2.8.2
tqdm>=4.62.3
click>=8.0.3
ollama>=0.1.5
